- updated to latest PrBoom SVN.

+++ prboom2/src/g_overflow.c
@@ -172,7 +172,7 @@ void InterceptsOverrun(int num_intercepts, intercept_t *intercept)
 
       InterceptsMemoryOverrun(location, intercept->frac);
       InterceptsMemoryOverrun(location + 4, intercept->isaline);
-      InterceptsMemoryOverrun(location + 8, (int) intercept->d.thing);
+      InterceptsMemoryOverrun(location + 8, (int)(intptr_t) intercept->d.thing);
     }
   }
 }
